+++ src/Model/Ssr/Template/Loader.php
@@ -4,14 +4,13 @@ declare(strict_types=1);
 
 namespace Omikron\Factfinder\Model\Ssr\Template;
 
-use Mustache_Loader;
-use Mustache_Source;
+use Handlebars\Loader as HandlebarsLoader;
 use Omikron\Factfinder\Api\Filter\FilterInterface;
 
-class Loader implements Mustache_Loader
+class Loader implements HandlebarsLoader
 {
     public function __construct(
-        private readonly Mustache_Loader $loader,
+        private readonly HandlebarsLoader $loader,
         private readonly FilterInterface $filter,
     ) {
     }
@@ -22,6 +21,7 @@ class Loader implements Mustache_Loader
     public function load($name)
     {
         $template = $this->loader->load($name);
-        return $template instanceof Mustache_Source ? $template : $this->filter->filterValue($template);
+
+        return $this->filter->filterValue($template);
     }
 }